Split out MatchingControlPaths to separate source file
